modified tex.vim to link instead of copy

+++ dotfiles/install.sh
@@ -44,8 +44,17 @@ fi
 # Link dotfiles
 link_files () 
 {
-    ln -s $1 $2
-    success "Linked $1 to $2"
+    info "Linking $1 to $2"
+    if [ -f $2 -o -d $2 ]
+    then
+        info "$2 exists. Replacing old version of $2 with new one"
+        rm -rf $2
+        ln -s $1 $2
+	success "Linked $1 to $2"
+    else
+        ln -s $1 $2
+	success "Linked $1 to $2"
+    fi
 }
 
 copy_files () 
@@ -61,8 +70,6 @@ copy_files ()
         cp -r $1 $2
 	success "Copied $1 to $2"
     fi
-
-    
 }
 
 symlink_dotfiles () 
@@ -90,7 +97,7 @@ copy_files $source $dest
 # Copy tex.vim (for vim-latex modifications)
 source="$HOME/Software/system-config/dotfiles/vim-latex/tex.vim"
 dest="$HOME/.vim/bundle/vim-latex-1.8.23/ftplugin/tex.vim"
-copy_files $source $dest
+link_files $source $dest
 
 # Link dotfiles functions
 symlink_dotfiles
